Allow to keep some or all logging info separately
* lisp/org-agenda.el (org-agenda-log-directory): New defcustom.
(org-agenda-get-day-entries-extern): New defun.
(org-agenda-list): Use `org-agenda-get-day-entries-extern' as an extra
source.
(org-agenda-format-item): Add a hack to show the original file name.

The idea is to have the following structure for logging info (clocks
and done timestamps for repeatable items):

Orglog/
└── 2017
    └── 2017-07-28.org

A single file would hold all the log info for the particular day.  All
agenda files would share it, each taking a first level headline with
the name of the file.